Se agregan validaciones para la clase interlocutor comercial, direccion, fecha

+++ InterlocutorComercial/Contrato.h
@@ -59,6 +59,6 @@ private:
 	Fecha fechaBajaContrato;
 	unsigned int codigoTarifa;
 	float consumoPromedio;
-	float posicionArchivo;
+	long posicionArchivo;
 	bool activo;
 };
